+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as Header } from './nodes/Header';
-export { default as Product } from './nodes/Product';
+export { default as Header } from './leafs/Header';
+export { default as Product } from './leafs/Product';
 export { default as ProductList } from './trees/ProductList';
