+++ src/components/trees/ProductList.tsx
@@ -1,6 +1,6 @@
-import type { ProductType } from "../../types";
-import styled from "styled-components";
-import Product from "../nodes/Product";
+import type { ProductType } from '../../types';
+import styled from 'styled-components';
+import Product from '../leafs/Product';
 
 interface ProductListProps {
   products: ProductType[];
